+++ lib/xiv_lodestone/lodestone_character.rb
@@ -39,14 +39,28 @@ module XIVLodestone
     end
 
     def initialise_profile(page)
+      @profile[:name] = Helper.get_name(page)
+      @profile[:server] = Helper.get_server(page)
+      @profile[:introduction] = Helper.get_introduction(page)
+      @profile[:title] = Helper.get_title(page)
+      @profile[:portrait] = Helper.get_portrait(page)
+      @profile[:hp] = Helper.get_hp(page)
+      @profile[:mp] = Helper.get_mp(page)
+      @profile[:tp] = Helper.get_tp(page)
+      @profile[:sex] = Helper.get_sex(page)
+      @profile[:race] = Helper.get_race(page)
+      @profile[:clan] = Helper.get_clan(page)
+      @profile[:nameday] = Helper.get_nameday(page)
+      @profile[:guardian] = Helper.get_guardian(page)
+      @profile[:city] = Helper.get_city(page)
+      @profile[:grand_company] = Helper.get_grand_company(page)
+      @profile[:free_company] = Helper.get_free_company(page)
       @profile[:disciple] = DiscipleList.new(page.xpath("//table[@class='class_list']/tr/td"))
       @profile[:gear] = GearList.new(page.xpath("(//div[@class='item_detail_box'])[position() < 13]"))
       @profile[:attribute] = AttributeList.new(page.xpath('//div[starts-with(@class, "param_left_area_inner")]/ul/li'))
       @mounts = MountList.new(page.xpath('(//div[@class="minion_box clearfix"])[1]/a'))
       @minions = MountList.new(page.xpath('(//div[@class="minion_box clearfix"])[2]/a'))
-      #@profile.merge!(parser.get_profile)
     end
-
     private :initialise_profile
   end
 end